Migrates stress test to ava
